Move functions dump and connect to new class Task.php
