+++ src/Webfactory/Slimdump/SlimdumpCommand.php
@@ -2,14 +2,11 @@
 
 namespace Webfactory\Slimdump;
 
-use Doctrine\DBAL\Connection;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Webfactory\Slimdump\Config\Config;
-use Webfactory\Slimdump\Config\ConfigBuilder;
-use Webfactory\Slimdump\Database\Dumper;
+
 
 class SlimdumpCommand extends Command
 {
@@ -32,69 +29,20 @@ class SlimdumpCommand extends Command
         ;
     }
 
+    /**
+     * @param InputInterface $input
+     * @param OutputInterface $output
+     * @return int|null|void
+     */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $dsn = $input->getArgument('dsn');
+        $config = $input->getArgument('config');
 
         if ($dsn === '-') {
             $dsn = getenv("MYSQL_DSN");
         }
 
-        $db = $this->connect($dsn);
-
-        $config = ConfigBuilder::createConfigurationFromConsecutiveFiles($input->getArgument('config'));
-        $this->dump($config, $db, $output);
+        $task = new Task($dsn, $config, $output);
     }
-
-    private function connect($dsn)
-    {
-        try {
-            return \Doctrine\DBAL\DriverManager::getConnection(
-                array('url' => $dsn, 'charset' => 'utf8', 'driverClass' => 'Doctrine\DBAL\Driver\PDOMySql\Driver')
-            );
-        } catch (Exception $e) {
-            $msg = "Database error: " . $e->getMessage();
-            fwrite(STDERR, "$msg\n");
-            exit(1);
-        }
-    }
-
-    /**
-     * @param Config $config
-     * @param Connection $db
-     * @param OutputInterface $output
-     * @throws \Doctrine\DBAL\DBALException
-     */
-    public function dump(Config $config, Connection $db, OutputInterface $output)
-    {
-        $dumper = new Dumper($output);
-        $dumper->exportAsUTF8();
-        $dumper->disableForeignKeys();
-
-        $platform = $db->getDatabasePlatform();
-
-        $fetchTablesResult = $db->query($platform->getListTablesSQL());
-
-        while ($tableName = $fetchTablesResult->fetchColumn(0)) {
-            $tableConfig = $config->findTable($tableName);
-
-            if (null === $tableConfig) {
-                continue;
-            }
-
-            if ($tableConfig->isSchemaDumpRequired()) {
-                $dumper->dumpSchema($tableName, $db, $tableConfig->keepAutoIncrement());
-
-                if ($tableConfig->isDataDumpRequired()) {
-                    $dumper->dumpData($tableName, $tableConfig, $db);
-                }
-
-                if ($tableConfig->isTriggerDumpRequired()) {
-                    $dumper->dumpTriggers($db, $tableName, $tableConfig->getDumpTriggersLevel());
-                }
-            }
-        }
-        $dumper->enableForeignKeys();
-    }
-
 }
